+++ NEWS.rst
@@ -5,17 +5,21 @@ Release History
 0.3 (2012-01-XX)
 ~~~~~~~~~~~~~~~~
 
+* Issue #3: ContextStack.register_exit() now accepts objects with __exit__
+  attributes in addition to accepting exit callbacks directly
 * Issue #1: Add ContextStack.preserve() to move all registered callbacks to
   a new ContextStack object
 * Wrapped callbacks now use functools.wraps to aid in introspection
 * Moved version number to a VERSION.txt file (read by both docs and setup.py)
 * Added NEWS.rst (and incorporated into documentation)
 
+
 0.2 (2011-12-15)
 ~~~~~~~~~~~~~~~~
 
 * Renamed CleanupManager to ContextStack (hopefully before anyone started
   using the module for anything, since I didn't alias the old name at all)
+
 
 0.1 (2011-12-13)
 ~~~~~~~~~~~~~~~~