+++ NEWS.rst
@@ -2,7 +2,7 @@ Release History
 ---------------
 
 
-0.3 (2012-01-XX)
+0.3 (2012-01-04)
 ~~~~~~~~~~~~~~~~
 
 * Issue #5: ContextStack.register no longer pointlessly returns the wrapped
@@ -13,7 +13,7 @@ Release History
 * Issue #1: Add ContextStack.preserve() to move all registered callbacks to
   a new ContextStack object
 * Wrapped callbacks now expose __wrapped__ (for direct callbacks) or __self__
-(for context manager methods) attributes to aid in introspection
+  (for context manager methods) attributes to aid in introspection
 * Moved version number to a VERSION.txt file (read by both docs and setup.py)
 * Added NEWS.rst (and incorporated into documentation)